Make new structure for nutrient stats report
